+++ api/src/services/graphql/schema/get-types.ts
@@ -9,8 +9,8 @@ import { GraphQLJSON, ObjectTypeComposer } from 'graphql-compose';
 import { mapKeys, pick } from 'lodash-es';
 import { GENERATE_SPECIAL } from '../../../constants.js';
 import { getGraphQLType } from '../../../utils/get-graphql-type.js';
-import { SYSTEM_DENY_LIST, type GQLScope } from '../index.js';
-import type { InconsistentFields, Schema } from './index.js';
+import { type GQLScope } from '../index.js';
+import { SYSTEM_DENY_LIST, type InconsistentFields, type Schema } from './index.js';
 
 /**
  * Construct an object of types for every collection, using the permitted fields per action type
