+++ api/src/services/graphql/schema/index.ts
@@ -10,7 +10,7 @@ import {
 import { fetchInconsistentFieldMap } from '../../../permissions/modules/fetch-inconsistent-field-map/fetch-inconsistent-field-map.js';
 import type { GraphQLParams } from '../../../types/index.js';
 import { reduceSchema } from '../../../utils/reduce-schema.js';
-import { GraphQLService, READ_ONLY, SYSTEM_DENY_LIST } from '../index.js';
+import { GraphQLService } from '../index.js';
 import { injectSystemResolvers } from '../resolvers/system.js';
 import { cache } from '../schema-cache.js';
 import { GraphQLVoid } from '../types/void.js';
@@ -32,6 +32,20 @@ export type CollectionTypes = {
 	UpdateCollectionTypes: Record<string, ObjectTypeComposer<any, any>>;
 	DeleteCollectionTypes: Record<string, ObjectTypeComposer<any, any>>;
 };
+
+/**
+ * These should be ignored in the context of GraphQL, and/or are replaced by a custom resolver (for non-standard structures)
+ */
+export const SYSTEM_DENY_LIST = [
+	'directus_collections',
+	'directus_fields',
+	'directus_relations',
+	'directus_migrations',
+	'directus_sessions',
+	'directus_extensions',
+];
+
+export const READ_ONLY = ['directus_activity', 'directus_revisions'];
 
 /**
  * Generate the GraphQL schema. Pulls from the schema information generated by the get-schema util.
